% Auto-generated: do not edit by hand
\name{filePicker}

\alias{filePicker}

\title{FilePicker component}

\description{

}

\usage{
filePicker(id=NULL, file_map=NULL, multiple_selection_allowed=NULL,
root_folder_id=NULL, selected_files=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{file_map}{Named list. JSON-style dictionary containing file tree.}

\item{multiple_selection_allowed}{Logical. Whether we are allowed to select multiple files}

\item{root_folder_id}{Character. The id of the root folder.}

\item{selected_files}{Unnamed list. Selected files.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

